Import liblognorm-1.0.1 as sysutils/liblognorm.
liblognorm shall help to make sense out of syslog data, or, actually,
any event data that is present in text form.

In short words, one will be able to throw arbitrary log message to
liblognorm, one at a time, and for each message it will output
well-defined name-value pairs and a set of tags describing the message.