Changes 5.3.28: Fixed tcl library linking for AIX 7.
Fixed a bug that could cause a trap if a lock timeout occurred while opening a database.

Fixed missing encryption support for the Android JDBC driver.

Fixed an incorrect message being displayed when the -l option was specified to the db_hotbackup utility.

Fixed a bug that DB_ENV->log_get_config did not work correctly before DB_ENV->open.

Fixed a bug that could cause a SQL build failure when FTS3 is enabled.

Fixed a bug that prevented in-memory SQL database from being created properly. They can now be created without the use of the SQLITE_OPEN_CREATE flag.

Fixed a memory leak in SQL online backup.

Added additional examples for C++

Fixed a bug that could make odbc fail to build with sqlite source code.

Fixed bugs in compaction of large keys in the upper levels of btrees.

The db utilities (db_xxxx) no longer operate on replication clients that are being automatically initialized. The DB_REP_LOCKOUT error is now returned.

Using DB_TXN_SNAPSHOT on an HA client will now result in an error.

Fixed a bug that prevented a sub-database from being created under the directory identified in DB->set_create_dir.

Fixed a race condition on a cursor when using a multi-threaded application with the SQL API.

Fixed a race condition in the failchk code when cleaning up mutexes.

Fixed a resource leak in the db->verify() function for btrees. The bug would slow down verification and possibly cause it to run out of memory.

Removed a potential hang when compacting databases with many duplicates.

Fixed a bug that caused a crash when reentering dbsql and specifying replication=on for a SQL database where replication was already enabled.

Fixed an incorrect recursive call dealing with joins.

Fix build failures in ado.net

Fixed a bug that could cause a JDBC build failure on recent versions of Visual Studio.

Correct build problems when building dbstl with gcc-4.7.3